Minor wording changes.
Changed "siloing" to "seperating" in "What it is not" section.
Some tweaks to second 3 of the COC to make it more relvant to a Slack group rather than an event:
Removed the word "verbal" from "Harassment includes: ..."
Updated refernce to "in slides" to be any documents uploaded or websites linked to.
Removed reference to "harassing photography or recording".
Added suggestion for how people may contact admins privately via DMs. This needs more thought I think.